Changed offset algorithm from calculating differences to set marks to differences in pasted text
BF: Saving position of deleted text and adding offset to that instead of jumping to mark and adding offset then (which doesn't work when the swap shortens the line and the mark now points to after the end of the line.
Added VIM7 custom operator.
Refactored code so that both the visual mode mapping and the operator use the same functions.