Adapt: Detection of deletion at the end of the line broken since Vim 8.2.0324
Since Vim 8.2.0324, the last change position can be one beyond the last [screen] column (col('$') == col("'.")); previously its value was reset to point at the character before the deletion if that happened at the end of the line (col('$') == col("'.") + 1).
The patch isn't about that, so it's not clear whether that change was intentional; as 2.5 years have passed already, let's take it at face value and implement a workaround here. # Please enter the commit message for your changes. Lines starting